pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "taro"
description = "An Example Project"
version = "0.1.0"
readme = "README.md"
dependencies = [
    "black",
    "numpy<2.3.0",
    "PyYAML==6.0.1",
    "yfinance",
    "psycopg2-binary",
    "sqlalchemy",
    "alembic",
    "python-dotenv",
    "flask",
    "bokeh",
    "bokeh_sampledata",
    "pandas",
    "gunicorn"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"taro.cli" = ["templates/*.html"]

[project.scripts]
taro = "taro.cli.__main__:main"
